+++ src/Rx.ts
@@ -54,8 +54,6 @@ import './add/operator/finally';
 import './add/operator/first';
 import './add/operator/groupBy';
 import './add/operator/ignoreElements';
-import './add/operator/inspect';
-import './add/operator/inspectTime';
 import './add/operator/every';
 import './add/operator/last';
 import './add/operator/map';
